+++ packages/core/src/modules/formula.ts
@@ -28,6 +28,8 @@ import {
   spillSortResult,
 } from ".";
 
+import { isLetterNumberPattern, removeLastSpan } from "../utils/index";
+
 let functionHTMLIndex = 0;
 let rangeIndexes: number[] = [];
 const operatorPriority: any = {
@@ -3272,6 +3274,19 @@ export function rangeSetValue(
   selected: any,
   fxInput?: HTMLDivElement | null
 ) {
+  const parser = new DOMParser();
+  const doc = parser.parseFromString(
+    `<div>${cellInput.innerHTML}</div>`,
+    "text/html"
+  );
+  const spans = doc.querySelectorAll("span");
+  const lastSpan = spans[spans.length - 1];
+
+  if (lastSpan && isLetterNumberPattern(lastSpan?.innerText)) {
+    const htmlR = removeLastSpan(cellInput.innerHTML);
+    cellInput!.innerHTML = `${htmlR}`;
+  }
+
   let $editor = cellInput;
   let $copyTo = fxInput;
   if (document.activeElement?.id === "luckysheet-functionbox-cell") {