+++ packages/react/src/components/SheetOverlay/InputBox.tsx
@@ -33,6 +33,15 @@ import ContentEditable from "./ContentEditable";
 import FormulaSearch from "./FormulaSearch";
 import FormulaHint from "./FormulaHint";
 import usePrevious from "../../hooks/usePrevious";
+import {
+  moveCursorToEnd,
+  isLetterNumberPattern,
+  removeLastSpan,
+  incrementColumn,
+  decrementColumn,
+  incrementRow,
+  decrementRow,
+} from "./helper";
 
 const InputBox: React.FC = () => {
   const { context, setContext, refs } = useContext(WorkbookContext);
@@ -42,11 +51,20 @@ const InputBox: React.FC = () => {
   const prevSheetId = usePrevious<string>(context.currentSheetId);
   const [isHidenRC, setIsHidenRC] = useState<boolean>(false);
   const [isInputBoxActive, setIsInputBoxActive] = useState(false);
+  const [activeCell, setActiveCell] = useState<string>("");
+  const [activeRefCell, setActiveRefCell] = useState<string>("");
   const [frozenPosition, setFrozenPosition] = useState({ left: 0, top: 0 });
   const firstSelection = context.luckysheet_select_save?.[0];
   const row_index = firstSelection?.row_focus!;
   const col_index = firstSelection?.column_focus!;
   const preText = useRef("");
+  const placeRef = useRef("");
+
+  useEffect(() => {
+    if (isInputBoxActive) {
+      setActiveCell(getCellAddress());
+    }
+  }, [isInputBoxActive]);
 
   const inputBoxStyle = useMemo(() => {
     if (firstSelection && context.luckysheetCellUpdate.length > 0) {
@@ -282,6 +300,84 @@ const InputBox: React.FC = () => {
       //   return;
       // }
 
+      /* Arrow navigation for cell reference starts here */
+      let allowListNavigation = true;
+
+      if (e.key === "Delete" || e.key === "Backspace") {
+        setTimeout(() => {
+          moveCursorToEnd(inputRef?.current!);
+        }, 5);
+      }
+
+      let refCell = placeRef.current;
+
+      if (e.key === "ArrowUp") {
+        refCell = decrementRow(placeRef.current);
+      } else if (e.key === "ArrowDown") {
+        refCell = incrementRow(placeRef.current);
+      } else if (e.key === "ArrowLeft") {
+        refCell = decrementColumn(placeRef.current);
+      } else if (e.key === "ArrowRight") {
+        refCell = incrementColumn(placeRef.current);
+      }
+
+      // current hack to for arrow navigation, try to find a better way like using rangeDrag
+      if (
+        e.key === "ArrowUp" ||
+        e.key === "ArrowDown" ||
+        e.key === "ArrowLeft" ||
+        e.key === "ArrowRight"
+      ) {
+        const parser = new DOMParser();
+        const doc = parser.parseFromString(
+          `<div>${inputRef?.current?.innerHTML}</div>`,
+          "text/html"
+        );
+        const spans = doc.querySelectorAll("span");
+        const lastSpan = spans[spans.length - 1];
+
+        const notFunctionInit = !document
+          .getElementById("luckysheet-rich-text-editor")
+          ?.innerText.includes("(");
+
+        // handling for inputbox active arrow navigation for cell reference input for functions like SUM(A1:A10)
+        const arrowRefNotAllowed =
+          lastSpan?.innerText.includes(")") ||
+          (notFunctionInit &&
+            /^[a-zA-Z]+$/.test(lastSpan?.innerText) &&
+            !_.includes(["="], lastSpan?.innerText));
+
+        if (
+          (lastSpan?.innerText === "(" ||
+            lastSpan?.innerText === "," ||
+            lastSpan?.innerText.includes(":") ||
+            lastSpan?.innerText !== ")") &&
+          !isLetterNumberPattern(lastSpan?.innerText) &&
+          !arrowRefNotAllowed
+        ) {
+          allowListNavigation = false;
+          inputRef.current!.innerHTML = `${
+            inputRef.current!.innerHTML
+          }<span class="fortune-formula-functionrange-cell" rangeindex="0" dir="auto" style="color:#c1232b;">${refCell}</span>`;
+
+          setTimeout(() => {
+            moveCursorToEnd(inputRef.current!);
+          }, 1);
+        }
+
+        if (isLetterNumberPattern(lastSpan?.innerText)) {
+          allowListNavigation = false;
+          const htmlR = removeLastSpan(inputRef?.current!.innerHTML);
+          inputRef.current!.innerHTML = `${htmlR}<span class="fortune-formula-functionrange-cell" rangeindex="0" dir="auto" style="color:#c1232b;">${refCell}</span>`;
+
+          moveCursorToEnd(inputRef.current!);
+          setTimeout(() => {
+            moveCursorToEnd(inputRef.current!);
+          }, 1);
+        }
+      }
+      /* Arrow navigation for cell reference ends here */
+
       if (e.key === "Escape" && context.luckysheetCellUpdate.length > 0) {
         setContext((draftCtx) => {
           cancelNormalSelected(draftCtx);
@@ -303,7 +399,8 @@ const InputBox: React.FC = () => {
         e.preventDefault();
       } else if (
         e.key === "ArrowUp" &&
-        context.luckysheetCellUpdate.length > 0
+        context.luckysheetCellUpdate.length > 0 &&
+        allowListNavigation
       ) {
         if (document.getElementById("luckysheet-formula-search-c")) {
           const formulaSearchContainer = document.getElementById(
@@ -336,7 +433,8 @@ const InputBox: React.FC = () => {
         e.preventDefault();
       } else if (
         e.key === "ArrowDown" &&
-        context.luckysheetCellUpdate.length > 0
+        context.luckysheetCellUpdate.length > 0 &&
+        allowListNavigation
       ) {
         if (document.getElementById("luckysheet-formula-search-c")) {
           const formulaSearchContainer = document.getElementById(
@@ -573,6 +671,15 @@ const InputBox: React.FC = () => {
     return `${columnChar}${rowNumber}`;
   }, [firstSelection]);
 
+  const wraperGetCell = () => {
+    const cell = getCellAddress();
+    placeRef.current = cell;
+    if (activeRefCell !== cell) {
+      setActiveRefCell(cell);
+    }
+    return activeCell || cell;
+  };
+
   return (
     <div
       className="luckysheet-input-box"
@@ -586,7 +693,7 @@ const InputBox: React.FC = () => {
           className="luckysheet-cell-address-indicator"
           style={getAddressIndicatorPosition()}
         >
-          {getCellAddress()}
+          {wraperGetCell()}
         </div>
       )}
       <div